codigo funcional, BD completa con las tablas, creado el DAO de cursoProfesor y su service, creado el DAO de Inscripcion y su service ( falta unir el service con la clase interfaz). PROBAR FUNCIONALIDAD Y CREAR LA UI PARA LO QUE FALTA.

+++ src/com/universidad/modelo/Curso.java
@@ -1,14 +1,14 @@
 package com.universidad.modelo;
 
 public class Curso {
-    private Integer id;
+    private Long id;
     private String nombre;
     private Programa programa;
     private Boolean activo;
     
     public Curso() {}
     
-    public Curso(Integer id, String nombre, Programa programa, Boolean activo) {
+    public Curso(Long id, String nombre, Programa programa, Boolean activo) {
         this.id = id;
         this.nombre = nombre;
         this.programa = programa;
@@ -16,8 +16,8 @@ public class Curso {
     }
     
     // Getters y Setters
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
     public Programa getPrograma() { return programa; }
